CV/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
timm>=0.5.4
numpy>=1.19.0 